Remove gate-requirements-install
It's redundant, which means it's more exposure to pypi, and it's the
simplistic version of something we have a more full version of.

The devstack gate tests test that changes to the requirements repo can
be installed in the context of the devstack install we do.

The pbr integration test installs the entire contents of the list
several times in the context of every project we have.

Change-Id: Ic15af3e965ee9c281bd8e9c62bd961cef032f640