add first pass recheck script
the follow script can be used to recheck reviews on a project that
haven't been touched in greater than a certain number of days. It's
the first step in creating a more generic tool for doing rechecks
on different projects automatically.

Updated to remove unused arg param (left over from optparse version)

Change-Id: I4d9182da9be2ea3396401452e9446c7761742804