Create new gates for openstack-manuals
Trunk and future stable branches will be checked by the new test.py
tool that implements four different checks.

The syntax and build ones should be run also as gate jobs. Syntax
is needed for checking of files that are not included in any books.

The niceness check is non-voting at the moment.

Change-Id: I45aff5a13e39a96134ed4a0353d15059e9e4af64