add level parameter to htmlify-screen-log
as an attempt to speed up both logstash, as well as user queries
of the logs, allow for a level parameter, which is the minimum
severity of a log line to display. This filter will work both with
html and text lines, and mean a lot less traffic over the wire.

Also provide links for making it easy to select log levels via
the HTML version of the logs.

Change-Id: I14a2912d15d25a613c90ef350bcd3a2350456c61