Fix missing branch angle line on first change
If the first change in a queue required an angled line to cap off
a branch line, it would not be added because the check to make
sure that it wasn't added to the first column was wrong.  Rework
it so that the "previous width" of the tree starts at 1 so that
if the width of the tree is more than 1 for the first change,
branch points are set for it.

Change-Id: I9aca034d0fe33db3bea5c58ed481010a1d2dd230