Set up new images on rackspace for check tests
Check tests queue is starved by the gate queue right now, so we want to
have a different pool of nodes for it. First of two patches, this one
sets up the images needed so that nodepool can get the image creation
done.

* modules/openstack_project/templates/nodepool/nodepool.yaml.erb:
Configure new devstack-precise-check image for use in check tests.

Change-Id: I8cff6f4f6cadcfa86d2488ae6768cefb1b2902a2