Adding puppet-vswitch project to Stackforge
Currently the puppet-vswitch module has many forks that are being
maintained by different groups. As the puppet-openstack modules,
particularly puppet-neutron/quantum and puppet-nova depend upon
this module, this update adds a new module to create a single
point for development and maintenance under the puppet-openstack
unbrella.

puppet-lint and puppet-syntax jobs are currently non-voting

This project targets both grizzly and havana, and only has
a master branch.

Change-Id: I1895bf05ebf15235fdfc57225a1fa5dda9724ef7