Gate devstack-gate changes on stable branch jobs
Since openstack-infra/devstack-gate only has a master branch,
changes to it did not previously get tested against stable branches
of devstack and tempest and the integrated projects. Correct this
asymmetry by adding versions of those jobs with appropriate
overrides forcing use of the target stable branches.

* .../jenkins_job_builder/config/devstack-gate.yaml: Add templates
for the pipeline-specific stable branch override jobs and put them
in the job group so that they'll get created.

* .../zuul/layout.yaml: Add the new jobs to the devstack-gate
project.

Change-Id: I487438cdde84b57d338893800e91bfccd6e49d0a