switch over to os-loganalyze on logs.o.o
this is the switch over for os-loganalyze on logs.o.o. It direct
deploys from git to the environment. It will bring with it
support for keystone, console logs, and being smarter about when
level filtering is applied.

While I would still like the have a tick / tock CD environment to
deploy to dev then to prod on a stable branch, there are enough
fixes in os-loganalyze over the old htmlify script that it's worth
deploying to master now, then work on CD next.

Change-Id: Iaa0c1993128032842b0d6ee496b0047fb65cb6a8