Check diff against existing req proposal changes.
* modules/jenkins/files/slave_scripts/propose_requirements_update.sh:
Previously this script updated requirements against tip of master and
attempted to push that to Gerrit. In the event where there is no diff
between requirements changes to tip of master and an existing
requirements proposal change the script will fail. Now if there is an
existing change check it out and use that as the base for the changes,
this allows us to check for diffs against the existing change.

Change-Id: I38958c1eafc12fb54a28f5d5ba22821b06cbe70a