Adding support to a generic zuul server in JJB for devstack-gate
This patch adds the support to a generic zuul server
in JBB for devstack-gate jobs, moving the
devstack-gate.yaml file from files to template
and adding the export ZUUL_URL=<= zuul_url> in each
job. This way, when devstack-vm-gate-wrap.sh is called
it will get the files from the specific zuul url.
If none is configured. The default http://zuul.openstack.org/p
will be used. It also changes the jenkins.pp file to
use devstack-gate.yaml as a template.

Change-Id: I6e3d70431098d0c3c6e4b133ff6ce295635409d0