Separate mailing list for StackForge projects
It has been suggested that a separate list for discussion of
projects hosted on Stackforge would be of benefit to reducting the
current list traffic on openstack-dev@lists.openstack.org. See:

http://lists.openstack.org/pipermail/openstack-dev/2013-November/019103.html

* modules/openstack_project/manifests/lists.pp: Add a new mailing
list at stackforge-dev@lists.openstack.org for use in discussing
StackForge projects as a preferred alternative to the existing
openstack-dev list.

Change-Id: I5dbb7b2e15c92636d86c91c0c128145b8684373f