Stop running propose-requirements-updates job
The propose-requirements-updates job runs from the post pipeline,
and zuul has no context there for which branch holds the corresponding
merged change. Because of this, the job proposes requirements
changes from stable branches to master, which is clearly not what we
want. Contain the damage by stopping it for now, and we can add it
back once it's properly branch-aware.

Change-Id: I8232184cbaac0980f7547e46c9cab6856c973499