Make python33 gate job for Rubick non-voting
While we must see the status of unittests for Rubick code under python 3.3,
we're still in early development stage and things tend to break from time
to time. So we want to make python33 job non-voting for the time being,
and turn it on again once we could figure out issues with external python
dependencies which are not 3.3 compliant.

Change-Id: If09de0b357dad876a33659a945806f9e07687d2d