Add gate-solum-devstack job
Add an initial devstack job for the stackforge project, solum.  It just
installs solum's devstack integration and runs devstack.  That just
ensures that the devstack integration doesn't blow up just trying to
start up the service(s).  There's not much else to test yet, but this is
a starting point to build on.

Start this job out as non-voting so that we can make sure it works as
expected before turning it on to be gating.

Change-Id: I4fd370abfcc51b71b250298163d1348a8a5c1bef