Add gate-manila-devstack job
Add an initial devstack job for the stackforge project, manila.
It just installs manila's devstack integration and runs devstack.
That just ensures that the devstack integration doesn't blow up
just trying to start up the service(s).

Start this job out as non-voting so that we can make sure it works as
expected before turning it on to be gating.

Change-Id: I7fde5a3d456a66404e86e7ead7e51038a291f039