Handle lack of rpms or debs in devstack-cache.
* modules/openstack_project/files/nodepool/scripts/devstack-cache.py:
When running on Ubuntu only debs will be listed for install, when
running on Fedora only RPMs will be listed for install. Handle the lack
of the other when checking which should be installed (use dict.get
instead of a normal index operation which will throw a KeyError instead
of returning None).

Change-Id: Iafb57f43ce6b51414593b909e1a9fde7c7c0ebf0