Add grenade-dsvm-forward job template
The stable grenade job used to be a place holder for the fact
that folsom => grizzly upgrade never was quite working, so we
ran grenade on the grizzly gates in a non voting way.

This is now replaced with the "forward" job, which is landing
a new change on the old side of the tree, and attempting to
upgrade forward. A push up upgrade (instead of a pull up
upgrade).

Take over the stable job for these purposes, putting it in non
voting mode for now to ensure that we're working before we light
it.

Change-Id: Iee18bf97a42dab3cdbef3066446b540e77072d73