Add grenade-vm-stable job template
The stable grenade job used to be a place holder for the fact
that folsom => grizzly upgrade never was quite working, so we
ran grenade on the grizzly gates in a non voting way.

We now have a new meaning for the stable grenade jobs, which
is forward upgrade compatibility. Thus doing the push up upgrade
(i.e. try to roll forward to havana on a grizzly update).

Take over the stable job for these purposes, putting it in non
voting mode for now to ensure that we're working before we light
it.

Change-Id: Iee18bf97a42dab3cdbef3066446b540e77072d73