use queries directory instead of file
This needs to land after I86317fcf6f1886ab5b6c0ee154b29e71865c52b7

moving elastic-recheck to a queries directory instead of a file
means we need to make the graphs support that. Also refactor out
the er_state_dir to make changing that easier when we pull in the
web console.

Change-Id: Ifcc71511c066e8aee5cb74e013555c628f270e29