Make job names shorter
Replace the use of '-default' in the new JJB devstack templates
with the empty string.

Shorten 'devstack-vm' to 'dsvm'.

Add 'devstack' in the position normally occupied by 'tempest' if
appropriate for the job (eg cells).

Set the single use flag for jobs that match 'dsvm'.

Add 'dsvm' to the requirements integration job name (which incidentally
was not matching the regex before, which is an error).

Adjust elastic-recheck graph and log host directory patterns to
match new job names.

Change-Id: I79562f79f41c02b5898a1c5e15194ed9d7dc1a39