Setup builds for openstack Gerrit ver 2.8
This patch sets up tooling and jobs required for building an openstack fork of
Gerrit ver 2.8.  Buck is the new build system for gerrit so we need to add
that toolchain.  Also buck does not have packages for easy installation.  It
needs to be built from source using ant.  This patch sets up two sets of jenkins
jobs to build gerrit, one for ver 2.4.x (maven) and one for ver 2.8.x (buck).

Change-Id: I1de9f127fa6826fb0a686d12dc77c0edff34be7c