Setup tooling and jobs for openstack Gerrit ver 2.8 and beyond
Previous versions of Gerrit used maven as the build tool, it has since switched
to use buck. I believe the switch happened in ver 2.7.  We are now hosting
a fork of Gerrit ver 2.8 so we need to setup tooling and jobs to build
newer versions of gerrit.  This patch sets up tooling and jobs required for
building the openstack Gerrit ver 2.8.  We now have two sets of jenkins jobs
to build gerrit, one for Gerrit prior to ver 2.4.x (maven) and one for Gerrit
post ver 2.8 (buck).  Buck does not have OS packages for easy installation
so it needs to be built from source using ant.

Partial-Bug: #1082781
Change-Id: I1de9f127fa6826fb0a686d12dc77c0edff34be7c