Setup tooling and jobs for openstack Gerrit ver 2.8 and beyond
Previous versions of Gerrit used maven as the build tool, it has since switched
to use buck. I believe the switch happened in ver 2.7.  We are now hosting
a fork of Gerrit ver 2.8 so we need to setup tooling and jobs to build
newer versions of gerrit.  This patch sets up tooling and jobs required for
building the openstack Gerrit ver 2.8.  We now have two sets of jenkins jobs
to build gerrit, one for Gerrit prior to ver 2.4.x (maven) and one for Gerrit
post ver 2.8 (buck).

Notes:
1. Buck does not have OS packages for easy installation so it needs to be built
   from source using ant.
2. Maven jobs will only run when commits are made to 2.4.x
3. The type of build to run is determined in the jenkins jobs.  Maven if there
   is a 'pom.xml' file and buck if there is a 'BUCK' file.
4. The maven jobs were converted to freestyle jobs so we could run a shell
   builder.

Partial-Bug: #1082781
Change-Id: I1de9f127fa6826fb0a686d12dc77c0edff34be7c