Ensure solum code is checked out
The solum devstack job has to install the devstack support for solum out
of the solum tree before running devstack.  This job assumed the solum
tree would be at the location where devstack puts it, but at this point,
devstack has not run yet.

Adding stackforge/solum to PROJECTS will get the devstack-gate script to
check it out before running this hook.

Change-Id: Ia682041e31f5e6783260d855c462df376a6d447c