Initial commit of a XenServer builder script
Copy this script to an Ubuntu 13.04 performance instance, and execute it
with the parameter "minvm". It will convert the instance to a XenServer. A
minimal precise VM will be accessible through the public IP. The XenServer
will be accessible through an internal virtual "mgt" network.

Staging VM username:
    root (use nova to inject ssh key)
XenServer's password:
    xspassword
XenServer's IP address:
    192.168.32.2

Flavor:
    performance1-8
Image:
    Ubuntu 13.04 (Raring Ringtail) (PVHVM beta)

Change-Id: Ibe94e56f6f3ac890e7f269042bc1db525b413efe