XenServer installer script
Related to XenServer/GateIntegration:
    https://wiki.openstack.org/wiki/XenServer/GateIntegration

This script will convert a Rackspace image to a XenServer 6.2 with an
appliance on top of it. Copy convert_node_to_xenserver.sh to a Rackspace
instance, created form the image 62df001e-87ee-407c-b042-6f4e13f5d7e1
and execute it with the first parameter being the desired password for
the XenServer, and the next parameter being a virtual appliance.

Exapmle:
    mkdir /opt/nodepool-scripts
    scp convert_node_to_xenserver.sh instance:/opt/nodepool-scripts/
    ssh instance /opt/nodepool-scripts/convert_node_to_xenserver.sh \
        password \
        http://downloads.vmd.citrix.com/OpenStack/\
            xenapi-in-the-cloud-appliances/1.0.0.xva

The incubator project for this script:
    https://github.com/citrix-openstack/xenapi-in-the-cloud

The project page for generating the appliance:
    https://github.com/citrix-openstack/openstack-xenapi-testing-xva

As the script is finished, the appliance will be accessible on the
instance's public IP. Dom0 could be reached from domU on the IP:
192.168.32.3

Flavor:
    performance1-8
Image:
    62df001e-87ee-407c-b042-6f4e13f5d7e1

Change-Id: Ibe94e56f6f3ac890e7f269042bc1db525b413efe