Use openstack-doc-tools for validation of operations-guide
This is the first step in using the repo openstack-doc-tools for
gating jobs. It starts with the operations-guide repository, in the
future openstack-manuals and api-site will also use openstack-doc-tools.
Using a single repository allows us sharing of the tools instead of
needing to copy them around.

This change creates a new build  openstack-doc-tools-checkout and uses
it for the gate-openstack-manuals-operations-guide.

Change-Id: I15e3e831a56e726b47f744900b8d6c410ab1875c