Add welcome_message.py to patchset-created trigger
Don't merge this yet :)

This patch adds the welcome message script to the patchset-created
trigger in gerrit. In theory it should post a lovely message to
any committer making a patchset for the first time.

However, it has not been tested, and review-dev is not available.
So, we first need to neuter welcome_message.py in jeepyb so it
won't cause catastrophic failures, and then merge this to test
in production, like a boss.

Change-Id: I4f04fcd2bd43f680f0d5032fc613194be8dd2586