Split config from projects list
The projects list is a common list for devs to interact with. The config
in the list is not, but the config in the list means the file needs to
be in an erb template.

Split the two concerns, similar to zuul. Put the config in a config file
and the project data in a yaml file.

Do it in two steps so that we can land this, then land a change to jeepyb
that consumes the config differently, then remove the config from the
yaml.

Change-Id: I708b8655b4b1ce377f3b7369e987418c1d72d977