zuul status: Collapse changes by default
Collapse change details by default.  Add a handy checkbox that you can
check that will expand all changes by default.

Once you click on the header of a change, it gets put in one of the
lists tracking state and is no longer affected by the checkbox that sets
the default.

Change-Id: I9055f7832f46d61477bc4812f34ca970c2166d00