Make cells job gating
We've had a cells job running for a while now and it  has been working
reliably.  Upgrade it to gating for the following projects:

  - nova, novaclient, gantt, ganttclient

Leave it alone for the rest.  It's still in some other check-only or
experimental queues.  Let's live with it gating nova for a while before
adding it further.

Change-Id: Ib483084bc2371089625caa4279a41c0992adfaef