Zuul layout: start using templates
This is a mostly no-op change where openstack projects that have pep8,
docs, python26 and python27 tests start using a template that provides
them.

Some projects will gain coverage or branch-tarball post jobs, which
should be harmless.

Many projects will have the order of tests in their reports changed,
which should also be harmless.

This is the first in a series of patches to replace lengthy job lists
with templates.

**Note** We require a production Zuul restart to upgrade before this
change can be merged.

Change-Id: I4de3fececb5c3f8013d9714b9ed405e08c822c91