Remove most periodic devstack-gate jobs
Remove the peridoc devsack-gate job definitions that are similar to
current gating jobs, using the pipeline parameter of the gate jobs
to produce jobs of the same name with nearly the same configuration.
Three periodic-only jobs remain; they will be folded into
devstack-gate.yaml in a future change.

Each job that is changed will change its log location and the zuul
branch override mechanism.  That is expected and normal.  A few
jobs will change their timeout values and collect new devstack-gate
options from the gate jobs (which seem reasonable and probably just
didn't make it over to the periodic variants in previous changes).

Change-Id: Id45d0c588c8e47a53b861e8440c2a367dfc89f43