Combine periodic grizzly and havana pipelines
Having two pipelines will cause twice the number of emails to be
sent.  Since the go to the same email list anyway, combine the
grizzly and havana periodic pipelines.  Keep the templates separate
to continue to support a project having one set of stable jobs but
not another.

Also, set the trigger minute to be different for each of the
periodic pipelines.  Zuul matches trigger events to pipelines based
on time specs, which means if 4 pipelines have timer triggers with
the same values, each pipeline is triggered 4 times (so 16 items
are enqueued across all 4 pipelines instead of 4 items).  This should
obviously be fixed in Zuul, but do this for now.

Add a customized subject to the emails sent by the pipeline.  This is
anticipating a change to Zuul, but is syntax-compatible with the
currently-running Zuul.

Change-Id: I1a8943af435115cb1aaf3c29fb2144e3be9e40bf