Add new javamelody-plugin project
This is a plugin that monitors gerrit using javamelody.

There currently is no repository for non-core plugins so there isn't a way to
consume the binaries.  We need this project to track upstream and build it
ourselves until gerrit core stands up a repo to hold non-core binaries.  Here
is the discussion thread..
https://groups.google.com/d/msg/repo-discuss/M5eVf40B1MU/bj0wUa-dZFEJ

Change-Id: I9345b3e022f35a0a5d8716cd3649cf146a152c43