Post pre-releases on pypi for marconiclient
Marconi client is not at the point of having stable releases, nor
full-featured ones. As part of Marconi's development, we need to have
small releases of the client that are aligned to the current API in
order to allow other projects to start implementing things based on
Marconi.

The patch enables pypi uploads for pre-releases.

Change-Id: Ia03877537a8f1f3a13b4f88e725ea26eef9d9420