make a dedicated page for gate status
this makes the gate status page a dedicated top level page
instead of mixing it with elastic recheck data. The existing
graphs are brought over as is.

This also reformats the way the header links work to be less error
prone (by not doing 2 statements per line), and using .push instead
of explicit array indexes.

Change-Id: I354f6cfd7105a1d9df0d0ae19bd68745336f29ed