Increase timeouts for jobs doing tempest runs
To reduce test load to get more stability in the gate, we're looking at
reducing the number of tests that tempest runs in parallel.  This will
an increase in the test runtime that is sometimes too close to the
current timeout.  Increase the timeout to make sure tempest gets enough
time to run.

For refernece, here is the change to run less tests in parallel:

  https://review.openstack.org/#/c/65805/

Change-Id: I5114501094e96b72efa33baa5f6eae304f8e393a