Add manuals-upstream-translation-update to api-site
The api-site repository, prior to being moved out of
openstack-manuals used to publish its content to tx for
translation. This patch re-starts that behaviour, by adding
the translation-update job to the api-site repo's config
in zuul.

The manuals-upstream-translation-update is the correct job
for this repo, since it contains XML files and will post
translatable resources to the openstack-manuals group on tx.

Change-Id: I1d713ab004162ef73c906ecbd288ca60611eb3b6