Add an experimental functional job for neutron.
It is often desirable to test Neutron's interaction with the
underling OS and without involving Tempest.  Unit tests do not
allow sudo invocations and are not guaranteed to run in an
environment configured with the necessary dependencies.  This
change adds a new sudo-capable functional job that executes
after devstack has configured the host.

Change-Id: I89058bf0e7f4d008c7f3403109a744f524db50c3