Record short_build_uuid in logstash/ElasticSearch
In elastic-recheck we want to be able to go from a log URL in
jenkins' gerrit comments to a single build_uuid. We already have the patch
number and revision, but there can be multiple builds for a single patch and
revision (think multiple rechecks or reverifies). The log URL contains a short
build_uuid but we cannot do substring matches on build_uuid in
ElasticSearch, so add short_build_uuid to ElasticSearch so we can use it
in elastic-recheck.

Change-Id: I091da52bd7a71b198250b327d1b7e63af3eb4a3c