Added artifact upload of storyboard.
Given that storyboard can be statically hosted, we have the ability to upload
the compiled artifacts to a location as a developer sanity check (much like
we do with docs.o.o). This patch adds the SCP task to upload (and host) a
copy of the storyboard webclient to static.o.o so we can look at/demo it.

Change-Id: I4abe7c9de243dacdf1a55c5e3846c6a7645f2937