Move o.o user creation to it's own manifest.
This patch creates a manifest for user creation that doesn't live in
base.  This is the first of several patches to abstract the
openstack_project module.  Moving user creation out to it's own
manifest will allow consumers of this module to optionally select
the user creation step.

This does not affect existing functionality.  It is a patch to
to add additional functionality later.

Change-Id: Ie96a28f1b05e4222dcc96b92357958a6b3276f96