Add NPM mirror
Sometimes, the intertubes get clogged, which causes NPM a headache when it
tries to resolve packages. This patch sets up an npm repository mirror that
can be used for our own testing infrastructure. Furthermore, this sets us
up to (at some point in the future) contribute packages back to the npm world.

Change-Id: I21b9653dfadfbb34f6e8158cce059a571c46d7ba