Make the alphabetized check self-gating
With putting the alphabetize script into slave_scripts, we run the risk
of bugs entering the script without being tested before they are rolled
live and become part of the gate. This just recently happened. Instead,
move it to the tools dir, which makes it easy for a dev to run, and also
which makes changes to the script self-gating.

Change-Id: Ia5e1870f84b0cbc0c6f172b6982263440af5575e