Enable PATH for exec functions
This patch adds a default search path (equal to setting $PATH) for
exec routines to consume.  This removes the need to set path per
command (unless the path is nonstandard) or to fully qualify path
to executables.

This change is non-impacting on current configuration, and is
provided for future expansion.

Change-Id: Ia18f248b1c4824cd99bfe36f633faab249871c75