prepare_devstack: install linux-generic-lts-saucy
One of the types of failures in the top gate bug is due to a kernel
crash in network namespace code.  This patch installs a newer kernel on
the devstack nodes.  It installs the backported kernel and its headers
that are made available for hardware enablement purposes.  This gets
from a 3.2.0 kernel to a 3.11.0 kernel.

The patch is written as a new node prep script.  This should allow us to
define a new node type that uses it and then create an experimental job
to use that node type.  That will allow us to make sure everything is
working before flipping the switch on using this kernel everywhere.

Change-Id: I9652b4183153687cc1e9a9a270ae0f78d8928620
Related-bug: #1254890