Make requirement update proposals more robust.
When proposing requirement update changes ensure that the proper branch
can be checked out if the branch isn't master. Additonally, don't git
review -d an existing change as it may have merge conflicts with the
target branch. Instead make changes again the target branch, push them
to the existing change as a new patchset, and ignore failures of git
review when there are no new changes.

This allows changes to the requirements files to happen out of band from
these automated change proposals.

Change-Id: I7b81b87a90639a66dce5502cded5e37d96abd1f7