Add new job to build the javamelody gerrit plugin
There currently is no repository for non-core gerrit plugins so there isn't a way to
consume the binaries[1].  We need to build the binary from upstream so that it is
available for us to install into our gerrit deployment.  This change adds a periodic
job to build the javamelody plugin.

[1] https://groups.google.com/d/msg/repo-discuss/M5eVf40B1MU/bj0wUa-dZFEJ

Partial-Bug: #1082781
Change-Id: Ibd915d26baff9d6c7ca376c55b66252f3061494f