turn on config drive for all jobs
we had turned off config drive because having it on meant we weren't
testing the metadata server. However, turning it off means that we're
doing key injection on guests.

It turns out that if you make certain nbd calls at the same time as
netns calls in the kernel, you get a kernel panic in Ubuntu's kernel.
This completely blocked the neutron team from doing any code for the
last 2 weeks.

Revert to config drive only for now. A metadata path that doesn't
need nbd based key injection needs to be sorted so we can test that,
and avoid the kernel bug.

Change-Id: I44e6422ff43859b36944c84dc8e692b96489be40