Add a testr specific builder
For those projects that use testr but also
use tox it is useful to provide a way to
specify the tox envlist to run with and to
also provide the same testr results copying
that the standard python-jobs get. To enable
this modify the run-unittests.sh to take in a
envlist (full-name) and then add a new builder
type that can be used instead of the tox job.

Closes-Bug: #1273850

Change-Id: I737d4d266cd78dd9f2f14a12ebe0ef7db8fe133f