Replace run-tox with run-unittests
Functionally these two are nearly equivalent except
one slightly alters the name of the virtualenv and
the other has less verbose output. It makes sense that
those who use tox venvs would also like to post testr
results (which was lacking in run-tox) so it appears
that this can be better handled by just unifying both
of the scripts.

Closes-Bug: #1273850

Change-Id: I737d4d266cd78dd9f2f14a12ebe0ef7db8fe133f