Extend unittest jenkins builder
For those projects that use testr/nose and
tox and subunit it is useful to provide a way to
specify the tox envlist to run with and to
also provide the same subunit results copying
that the standard python-jobs get. To enable
this modify the run-unittests.sh to take in a
envlist (full-name) and then add a new builder
type that can be used instead of the tox job.

This gains the job users the following:

- Use the subunit output and the posting of
  this data to there job results (which was
  not in the tox script).

Closes-Bug: #1273850

Change-Id: I737d4d266cd78dd9f2f14a12ebe0ef7db8fe133f