Module to handle package management via hiera
This module adds future extensibility to the o.o puppet framework
by allowing us to put package install data into hiera.

This would necessitate adding a public hiera data folder, and
adjusting the hiera.yaml to search both public and private.  If
used, it would help resolve issues where a package might need to be
installed for several components, but not all components might be
present on a given node.

It allows bulk package management, using [installed, latest, removed]
as well as detailed single package management.  When dealing with
bulk changes, it checks for an existing definition before attempting
the action, and outputs a warning to the log if required.

Hiera data example follows. (Note, this is from the included readme)
This module also requires the deep_merge gem to be installed.

packages::install:
  - 'curl'
  - 'unzip'

packages::latest:
  - 'puppet'
  - 'hiera'

packages::remove:
  - 'nethack'

packages::versioned
  gear:
    ensure: 'latest'
    provider: 'pip'
  deep_merge:
    ensure: 'latest'
    provider: 'gem'

Change-Id: Icf61a3ed86f59f26f8ab0cb2d95f82ed2db19b63