fix some small syntax issues in puppet manifest.
1. Template openstack_project/mavencentral-curl.erb uses options
mavencentral_username and mavencentral_password, and site.pp uses these
parameters from hiera in node pypi.slave.openstack.org but the class
does not support the arguments explicitly.  Arguments supported by the
class should be called out so that users of the class are aware of the
refrenced variables.

2. status.pp update provides escaping \$ to avoid possible confusion
down the line that this should not be expanded by puppet, and allows
syntax checkers like geppetto to not trigger errors when checking
strings for this.   Same command line output is still formed with
explicit escape for $ which should not be expanded.

Change-Id: I776748487671338d524a6b892e73796336a37491