Revoke sudo permissions before infra jobs
This should actually be a noop -- the sudo file should not be in
place yet, but the revoke-sudo builder needs to work regardless
of whether the file is in place or not.  This tests the latter
condition.

Next step is to add this to all jobs that run on bare-precise nodes
and then enable sudo on bare-precise nodes.

Change-Id: Ib6241d39cefc2a4df66e6f1b3f2e9e18bd08db09