Add the kite (KDS) project and kiteclient project to gerrit/jenkins
Add the new repositories for Kite (KDS) and kiteclient to gerrit
and jenkins. This project is under the OpenStack Identity Program and
provides a service for symmetrical keys to the cloud services (e.g. Nova)
for secure messaging. Currently the project lives within the Keystone
repository and should be split out into it's own repository to ensure
there are no conflicts and a clean delineation between the code bases.

A new set of core groups are added for kite specifically, since the
core teams for each project are likely to diverge, however, the
PTL of the Identity Program should be the same unless Kite is re-homed
under another program.

Change-Id: I0f4cf0b343d956e83e2ed85b44e03b6e3e1e5873