Re-enable requiring a recent check result for gate
This corrects the error that was causing changes not to be
enqueued in the gate after a verified +1 from jenkins.

If jenkins leaves a -2 vote, allow either a "reverify" comment or
an approval vote to retrigger the check pipeline.  This should then
retrigger the gate pipeline if it leaves a +1 verified vote as a
result.

Change-Id: I8ba26f254a1077b274182cb3212962428b00b478