add gerrit-git-prep option to clone into a sub directory
Sometimes it may be necessary to clone code into a sub directory.  I've noticed this
when I was trying to setup a jenkins job for gerrit[1].  In that job I needed two git
clones and they needed to be at the same directory level. A workaround could be used where
you copy into sub directory after clone but I found that not very clean.  Extending
gerrit-git-prep to clone into a sub directory will make the jenkins jobs cleaner and more
flexible.

[1] I1de9f127fa6826fb0a686d12dc77c0edff34be7c

Change-Id: I46cca9619828241de177417c0f9194987e436328