Run all non-sensitive jobs on single-use workers
* modules/openstack_project/files/zuul/layout.yaml: Get rid of the
transitional matches for setting the single use parameter function
on specific job names and just set it on everything as a default.
Add patterns for jobs which need long-running slaves to apply a
reusable node parameter function instead.

* modules/openstack_project/files/zuul/openstack_functions.py: Add a
parameter function for reusable nodes which toggles the offline
behavior in the opposite direction from single use.

Change-Id: I2a90c5b120a50bb8033c891702185f054ee84d35