Fix pip on py3k/pypy nodes
The puppet modules that run after install_puppet blow away the pip
installed pip and replace it with distro pip and setuptools. Those
are an abomination and cause various builds to break.

After setting up the script, remove the python-setuptools and python-pip
packages and re-run the commands to get pip and setuptools on the system.

Change-Id: Ie862cb35dd93cb655941ac556d2b07d5e6be9ca1