Nodepool should clone the config repo from git.o.o
* .../nodepool/scripts/prepare_node.sh: Switch the config clone
operation from the Gerrit server to the Git server farm, since the
former does not support git protocol. Also make it a shallow clone
since we just need the files. While at it, switch to retrieving the
bootstrap script from with plain HTTP because we're not retrieving
other things via SSL either.

Change-Id: I1dc0697bb3f9832d180b6b1ebffaa563bd886281